Ajuste editar campo select

+++ app/Http/Controllers/CrearController.php
@@ -135,14 +135,17 @@ class CrearController extends Controller
         $productos = Productos::findOrFail($id);
         $idProducto = $productos->id;        
         
+        $idVenta = '';        
         $ventas = Ventas::where("productos_id", $idProducto)->get(); 
                 
         foreach ($ventas as $ventaRow){
             $idVenta = $ventaRow->productos_id;
         }
 
-        DB::table('ventas')->where('productos_id', '=', $idVenta)->delete();
-
+        if($idVenta != ''){
+            DB::table('ventas')->where('productos_id', '=', $idVenta)->delete();    
+        }
+        
         $productos->delete();
 
         return redirect()->route('home');
